backend/requirements.txt
Django==4.2.10
django-colorfield==0.11.0
django-filter==24.1
djangorestframework==3.14.0
djoser==2.2.2
drf_extra_fields==3.7.0
gunicorn==21.2.0
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
